[ORM] [IDS] [!LONG] Enabling non-Long @Id fields. - @Id must be Comparable - Yopable.setId and Yopable.getId now support Comparable and still can be overriden. - There is still a Yopable.setId(Long) : most Yopable should have a Long ID and this keeps the Test objects compatible.

+++ src/test/java/org/yop/orm/simple/CountTest.java
@@ -82,7 +82,7 @@ public class CountTest extends DBMSSwitch {
 			Assert.assertEquals(0, count);
 
 			for (Pojo pojo : pojos) {
-				if (pojo.getId() % 3 == 0) {
+				if ((long) pojo.getId() % 3 == 0) {
 					pojo.setActive(true);
 					Jopo jopo = new Jopo();
 					jopo.setName("jopo #1 for Pojo #" + pojo.getId());
